add all files new

+++ database/migrations/2025_07_07_062424_create_student_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('mobile');
+            $table->bigInteger('mobile');
             $table->string('class')->nullable();
             $table->string('session')->nullable();
             $table->timestamps();
